requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
numba>=0.58.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.0.0
Pillow>=10.0.0
scipy>=1.11.0
